caps/eth.md: add eth/63 protocol specification (#62)
This spec was adapted from the content of the Ethereum wiki. The spec is
reformatted to include just the most recent version as well as a change
log listing past versions and their changes.

Sections explaining protocol semantics were expanded to cover block
propagation, full sync and fast sync as implemented in geth.